+++ snowchains_core/CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## [Unreleased]
 
+### Added
+
 - Enabled downloading all of the test cases on Dropbox.
+
+### Changed
+
+- Added `url` and `display_name` fields to `RetrieveTestCasesOutcomeContest`.
+- Made `Outcome`s `#[non_exhaustive]`.
 
 ## [0.4.3] - 2020-08-23Z